Fixed ambiguous call.

+++ Types.cpp
@@ -37,7 +37,7 @@ float int2::dot(float2 &other) const {
     return x*other.x+y*other.y;
 }
 float int2::distance(int2 &other) const {
-    return sqrt((x-other.x)*(x-other.x)+(y-other.y)*(y-other.y));
+    return sqrt((float)(x-other.x)*(x-other.x)+(y-other.y)*(y-other.y));
 }
 float int2::dot(int2 &other) const {
     return x*other.x+y*other.y;
@@ -51,7 +51,7 @@ float int3::dot(float3 &other) const {
     return x*other.x+y*other.y+z*other.z;
 }
 float int3::distance(int3 &other) const {
-    return sqrt((x-other.x)*(x-other.x)+(y-other.y)*(y-other.y)+(z-other.z)*(z-other.z));
+    return sqrt((float)(x-other.x)*(x-other.x)+(y-other.y)*(y-other.y)+(z-other.z)*(z-other.z));
 }
 float int3::dot(int3 &other) const {
     return x*other.x+y*other.y+z*other.z;
